# api/openapi.yaml
openapi: 3.1.0
info:
  title: Verbum API
  version: 0.1.0
  description: |
    The contract between the Verbum apps (iOS, Android) and the backend (docs/PRODUCT.md §45).
    The apps' client interfaces — `GraphClient`, `SearchClient`, `ContextClient`, `TimelineClient`,
    `DailyVerses` — were written first; this document is their wire shape, and `examples/` holds
    the responses both apps decode in their contract tests. Change the JSON here, and the app tests
    break before a user does.

    Conventions
    - Editorial GET routes are public and cacheable. Paid POST routes require a Firebase ID token.
    - Public search is lexical; a verified identity enables paid semantic retrieval.
    - Scripture text is NOT served here: the apps read it from bible.helloao.org (and the bundled WEB).
      References are structured (`PassageReference`), never prose, so the apps render them in the
      reader's translation and language.
    - Every relationship, detail and dating claim carries `sourceReferenceIds` (§33). Content only
      reaches this API through the editorial pipeline with human review (§32).
    - Errors are one shape (`Problem`) with a small `code` vocabulary (§52); the apps map codes to
      states and never show `message` to users.
    - Ids are opaque strings. Fixture ids look like `fixture.person.david`; production ids may differ.
security: []   # editorial content stays public; paid operations override this below
servers:
  - url: https://api.vendlydigital.com.br
...
  - url: http://localhost:8080
    description: local backend
...
paths:
...
  /v1/entities:
    get:
      tags: [entities]
      operationId: listEntities
      summary: Every entity of one kind, for the Explore lists (§7). Passage nodes are never listed.
      description: Backs `GraphClient.entities(type)`. Sorted by name in the requested language.
      parameters:
        - name: type
          in: query
          required: true
          schema: { $ref: '#/components/schemas/BibleEntityType' }
        - $ref: '#/components/parameters/Lang'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [entities]
                properties:
                  entities:
                    type: array
                    items: { $ref: '#/components/schemas/BibleEntity' }
              examples:
                people: { externalValue: examples/entities/people.json }
...
components:
  securitySchemes:
    VerbumVoiceTicket:
      type: http
      scheme: bearer
      description: One-use ticket returned by createRealtimeSession, never a provider key.
    FirebaseIDToken:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Firebase SDK ID token for a registered or anonymous user, validated against the configured project including revocation.
  parameters:
...
    Lang:
      name: lang
      in: query
      required: false
      schema: { type: string, enum: [en, pt, pt-BR], default: en }
      description: Presentation language; pt aliases pt-BR. Missing translations fall back to en. Original scripts and identifiers remain unchanged. Unsupported values use en.
...
  schemas:
...
    BibleEntityType:
      type: string
      enum: [person, place, event, theme, passage, book, prophecy, originalTerm, historicalPeriod]
    BibleEntity:
      type: object
      required: [id, type, name]
      properties:
        id: { type: string }
        type: { $ref: '#/components/schemas/BibleEntityType' }
        name: { type: string }
        summary: { type: [string, 'null'], description: Short and non-interpretive (§3.5). }
        nameLanguage:
          type: string
          enum: [en, pt-BR]
          description: Selected presentation record, including fallback. OriginalTerm names keep the original script; its language is lexical.language. Absent for legacy entities without localization records.

    # ---- §22.2
...